+++ validate/email/src/index.ts
@@ -1,0 +1,1 @@
+export { validateEmail } from './validate.email.plugin'
