--- validators/email/metadata.json
@@ -1,70 +1,0 @@
-{
-  "timestamp": "2024-09-24T06-57-54-540Z",
-  "task": "Create an email validation Flatfile Listener plugin:\n      - Implement a RecordHook for email field validation\n      - Use regex patterns to check for valid email structures\n      - Verify domain existence using DNS lookups\n      - Implement disposable email address detection\n      - Add error messages for invalid email addresses\n      - Give the user reasonable config options to specify the Sheet Slug, the Field(s) that are the email(s), whether the validation should be done automatically",
-  "summary": "This is a Flatfile Listener plugin for email validation. It uses recordHook to validate email fields in a Flatfile sheet. The plugin checks for valid email format using regex, verifies the domain against a list of disposable email domains, and optionally performs a DNS lookup for domain validation. It includes configuration options for sheet slug, email fields, and automatic validation.",
-  "steps": [
-    [
-      "Retrieve information about Flatfile Listeners and RecordHook implementation.\n",
-      "#E1",
-      "PineconeAssistant",
-      "Provide an example of a Flatfile Listener using RecordHook for field validation",
-      "Plan: Retrieve information about Flatfile Listeners and RecordHook implementation.\n#E1 = PineconeAssistant[Provide an example of a Flatfile Listener using RecordHook for field validation]"
-    ],
-    [
-      "Get a regex pattern for email validation.\n",
-      "#E2",
-      "Google",
-      "regex pattern for email validation javascript",
-      "Plan: Get a regex pattern for email validation.\n#E2 = Google[regex pattern for email validation javascript]"
-    ],
-    [
-      "Retrieve information about DNS lookup implementation in JavaScript.\n",
-      "#E3",
-      "Google",
-      "how to perform DNS lookup in JavaScript",
-      "Plan: Retrieve information about DNS lookup implementation in JavaScript.\n#E3 = Google[how to perform DNS lookup in JavaScript]"
-    ],
-    [
-      "Get a list of common disposable email domains.\n",
-      "#E4",
-      "Google",
-      "list of disposable email domains",
-      "Plan: Get a list of common disposable email domains.\n#E4 = Google[list of disposable email domains]"
-    ],
-    [
-      "Create the email validation Flatfile Listener plugin using the gathered information.\n",
-      "#E5",
-      "LLM",
-      "Create a Flatfile Listener plugin for email validation using RecordHook. Use the regex pattern from #E2 for initial validation, implement DNS lookup using the method from #E3, and check against the disposable email list from #E4. Include config options for Sheet Slug, email field(s), and automatic validation. Use the structure from #E1 as a reference. Provide detailed error messages for invalid emails.",
-      "Plan: Create the email validation Flatfile Listener plugin using the gathered information.\n#E5 = LLM[Create a Flatfile Listener plugin for email validation using RecordHook. Use the regex pattern from #E2 for initial validation, implement DNS lookup using the method from #E3, and check against the disposable email list from #E4. Include config options for Sheet Slug, email field(s), and automatic validation. Use the structure from #E1 as a reference. Provide detailed error messages for invalid emails.]"
-    ],
-    [
-      "Verify the created plugin for correctness and completeness.\n",
-      "#E6",
-      "PineconeAssistant",
-      "Verify the following Flatfile Listener plugin for email validation: #E5. Check if it implements all required features, uses valid Event Topics, and includes the necessary config options.",
-      "Plan: Verify the created plugin for correctness and completeness.\n#E6 = PineconeAssistant[Verify the following Flatfile Listener plugin for email validation: #E5. Check if it implements all required features, uses valid Event Topics, and includes the necessary config options.]"
-    ],
-    [
-      "Make any necessary corrections based on the verification results.\n",
-      "#E7",
-      "LLM",
-      "Review the verification results from #E6 and make any necessary corrections or improvements to the email validation Flatfile Listener plugin.",
-      "Plan: Make any necessary corrections based on the verification results.\n#E7 = LLM[Review the verification results from #E6 and make any necessary corrections or improvements to the email validation Flatfile Listener plugin.]"
-    ],
-    [
-      "Perform a final check for unused imports and validate params and Event Topics.\n",
-      "#E8",
-      "LLM",
-      "Review the final version of the email validation Flatfile Listener plugin from #E7. Remove any unused imports, validate that all params are correct for the plugins used, ensure the code is valid, and confirm that the listener only subscribes to valid Event Topics.",
-      "Plan: Perform a final check for unused imports and validate params and Event Topics.\n#E8 = LLM[Review the final version of the email validation Flatfile Listener plugin from #E7. Remove any unused imports, validate that all params are correct for the plugins used, ensure the code is valid, and confirm that the listener only subscribes to valid Event Topics.]"
-    ]
-  ],
-  "metrics": {
-    "tokens": {
-      "plan": 2733,
-      "state": 3647,
-      "total": 6380
-    }
-  }
-}
