--- validators/email/src/disposable-domains.ts
@@ -1,1 +1,0 @@
-export const disposableDomains = ['asdf.com']
